+++ niveshpy/cli/models/__init__.py
@@ -1,1 +1,0 @@
-"""Module for CLI data models."""