+++ niveshpy/infrastructure/sqlite/models.py
@@ -51,7 +51,7 @@ class Account(SQLModel, table=True):
             name=self.name,
             institution=self.institution,
             properties=self.properties,
-            created_at=self.created_at,
+            created=self.created_at,
         )